+++ src/app/workers/offscreen-decoder.worker.ts
@@ -15,9 +15,6 @@ import { XpraVideoDecoder, XpraVideoDecoderLoader } from '../decoders/video-deco
  * Worker for offscreen decoding.
  */
 
-// importScripts("./lib/lz4.js");
-// importScripts("./lib/broadway/Decoder.js");
-
 // WindowDecoder for each window we have control over:
 const window_decoders = new Map<string, WindowDecoder>();
 
@@ -35,9 +32,7 @@ const image_coding = [
 const video_coding: string[] = [];
 if (XpraVideoDecoderLoader.hasNativeDecoder()) {
   // We can support native H264 & VP8 decoding
-  video_coding.push("h264");
-  video_coding.push("vp8");
-  video_coding.push("vp9");
+  video_coding.push("h264", "vp8");
 } else {
   console.warn(
     "Offscreen decoding is available for images only. Please consider using Google Chrome 94+ in a secure (SSL or localhost) context for h264 offscreen decoding support."
@@ -46,7 +41,6 @@ if (XpraVideoDecoderLoader.hasNativeDecoder()) {
 
 const all_encodings = new Set([
   "void",
-  "scroll",
   ...image_coding,
   ...video_coding,
 ]);
@@ -132,23 +126,19 @@ class WindowDecoder {
     } else if (image_coding.includes(coding)) {
       await this.image_decoder.convertToBitmap(packet);
     } else if (video_coding.includes(coding)) {
-      if (coding == "vp9") {
-        // Prepare the VP9 codec (if needed)
-        const csc = packet[10]["csc"];
-        this.video_decoder.prepareVP9params(csc);
-      }
-
       if (!this.video_decoder.initialized) {
         this.video_decoder.init(coding);
       }
-      await this.video_decoder.queue_frame(packet);
+      packet = await this.video_decoder.queue_frame(packet).catch((error) => {
+        this.decode_error(packet, error);
+      });
     } else {
       this.decode_error(packet, `unsupported encoding: '${coding}'`);
     }
 
     // Hold throttle packages for 500 ms to prevent flooding of the VideoDecoder
     if (packet[6] == "throttle") {
-      await new Promise(r => setTimeout(r, 500));
+      await new Promise((r) => setTimeout(r, 500));
     }
 
     // Fake packet to send back
@@ -180,7 +170,8 @@ class WindowDecoder {
           w: packet[4],
           h: packet[5],
         },
-        [packet[7]]
+        // Scroll does not hold a transferable type
+        coding == "scroll" ? [] : [packet[7]]
       );
     }
   }